The box (big) in multiple layers, for Raspberry pi
Large version of the box, with room for a full B+ version raspberry pi
in it. For ease of use and developpement, the pi 3 is recommended, since
it gets rid of the need for an external USB wifi.
However, if needs be, the pi can be installed with USB ports towards the
inside of the box, so that the wifi key doesn't stick out.